+++ core/modules/sources/synthesis/pluck.rs
@@ -6,7 +6,7 @@ pub struct Pluck {
     detune: f32,
     spread: f32,
     glide: f32,
-    dropdown: u32,
+    dropdown: usize,
 }
 
 pub struct PluckVoice {
